🗃️ Committing everything that changed 🗃️

+++ etc/install_plugins.sh
@@ -22,7 +22,7 @@ RUN_USER="${SUDO_USER:-$USER}"
 SRC_DIR="${BASH_SOURCE%/*}"
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 # Set bash options
-if [[ "$1" == "--debug" ]]; then shift 1 && set -xo pipefail && export SCRIPT_OPTS="--debug" && export _DEBUG="on"; fi
+if [ "$1" = "--debug" ]; then shift 1 && set -xo pipefail && export SCRIPT_OPTS="--debug" && export _DEBUG="on"; fi
 
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 TMUX_HOME="${TMUX_HOME:-$HOME/.config/tmux}"
@@ -30,17 +30,29 @@ TMUX_SHARE_DIR="${TMUX_SHARE_DIR:-$HOME/.local/share/tmux}"
 TMUX_PLUGIN_MANAGER_PATH="${TMUX_PLUGIN_MANAGER_PATH:-$TMUX_SHARE_DIR/tpm}"
 export TMUX_HOME TMUX_PLUGIN_MANAGER_PATH TMUX_SHARE_DIR
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-[[ -d "$TMUX_PLUGIN_MANAGER_PATH/.git" ]] && git -C "$TMUX_PLUGIN_MANAGER_PATH" pull ||
-  git clone "https://github.com/tmux-plugins/tpm" "$TMUX_PLUGIN_MANAGER_PATH"
+if [ -d "$TMUX_PLUGIN_MANAGER_PATH/.git" ]; then
+  git -C "$TMUX_PLUGIN_MANAGER_PATH" pull -qq &>/dev/null
+else
+  [ -d "$TMUX_PLUGIN_MANAGER_PATH" ] && rm -Rf "$TMUX_PLUGIN_MANAGER_PATH"
+  git clone "https://github.com/tmux-plugins/tpm" "$TMUX_PLUGIN_MANAGER_PATH" &>/dev/null
+fi
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-[ -x "$TMUX_PLUGIN_MANAGER_PATH/bin/install_plugins" ] &&
+if [ -x "$TMUX_PLUGIN_MANAGER_PATH/bin/install_plugins" ]; then
   "$TMUX_PLUGIN_MANAGER_PATH/bin/install_plugins"
+fi
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 mkdir -p "$TMUX_SHARE_DIR/resurrect"
-[[ -f "$TMUX_HOME/resurrect" ]] &&
-  ln -sf "$TMUX_HOME/resurrect" "$TMUX_SHARE_DIR/resurrect/last"
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-[[ -f "$TMUX_PLUGIN_MANAGER_PATH/tpm" ]] && [[ -f "$TMUX_SHARE_DIR/resurrect/last" ]] &&
+if [ -f "$TMUX_HOME/resurrect" ]; then
+  ln -sf "$TMUX_HOME/resurrect" "$TMUX_SHARE_DIR/resurrect/last"
+fi
+# - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+if [ -f "$TMUX_PLUGIN_MANAGER_PATH/tpm" ] && [ -f "$TMUX_SHARE_DIR/resurrect/last" ]; then
   echo "Install completed"
+  exitCode=0
+else
+  echo "Install failed"
+  exitCode=1
+fi
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 exit ${exitCode:-$?}
